requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
ok-script==1.0.26
    # via -r requirements.in
onnxocr-ppocrv4==0.0.5
    # via -r requirements.in
pyappify==1.0.2
    # via ok-script
certifi==2025.7.14
    # via requests
charset-normalizer==3.4.2
    # via requests
comtypes==1.4.11
    # via pycaw
darkdetect==0.8.0
    # via pyside6-fluent-widgets
idna==3.10
    # via requests
numpy==2.2.6
    # via
    #   opencv-python
    #   openvino
    #   shapely

opencv-python==4.12.0.88
    # via -r requirements.in
openvino==2025.2.0
    # via -r requirements.in
openvino-telemetry==2025.2.0
    # via openvino
packaging==25.0
    # via openvino
pillow==11.3.0
    # via onnxocr-ppocrv4
psutil==7.0.0
    # via
    #   ok-script
    #   pycaw
pycaw==20240210
    # via ok-script
pyclipper==1.3.0.post6
    # via onnxocr-ppocrv4
pydirectinput==1.0.4
    # via ok-script
pyside6==6.9.1
    # via pyside6-fluent-widgets
pyside6-addons==6.9.1
    # via pyside6
pyside6-essentials==6.9.1
    # via
    #   pyside6
    #   pyside6-addons
pyside6-fluent-widgets==1.8.3
    # via ok-script
pysidesix-frameless-window==0.7.3
    # via pyside6-fluent-widgets
pywin32==311
    # via
    #   ok-script
    #   pysidesix-frameless-window
requests==2.32.4
    # via ok-script
shapely==2.1.1
    # via onnxocr-ppocrv4
shiboken6==6.9.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
typing-extensions==4.14.1
    # via ok-script
urllib3==2.5.0
    # via requests
adbutils==2.11.0
#如果支持模拟器则添加这个
